finished new blackjack
